app/requirements.txt
fastapi
openai
python-dotenv
uvicorn[standard]
pytest
matplotlib
seaborn
black
flake8